third_party/blink/web_tests/platform/generic/external/wpt/html/editing/dnd/the-dropzone-attribute/dropzone_attribute-expected.txt

This is a testharness.js-based test.
PASS div.dropzone should not be undefined if it's been set
FAIL div.dropzone should not be null assert_not_equals: div.dropzone should not be null got disallowed value null
FAIL div dropzone idl attribute must reflect the content attribute of the same name assert_equals: div dropzone idl attribute must reflect the content attribute of the same name expected (string) "copy file:image/png file:image/gif file:image/jpeg" but got (object) null
PASS div dropzone content attribute is set to the literal value
Harness: the test ran to completion.
